% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vebcMeta.R
\name{generate_obm_metadata}
\alias{generate_obm_metadata}
\title{Generate OBM metadata Excel templates}
\usage{
generate_obm_metadata(
  table_name,
  schema_name = "sex_ratio_evolution",
  url = "https://openbiomaps.org",
  project = "sex_ratio_evolution"
)
}
\arguments{
\item{table_name}{Character. The name of the table to query from OpenBioMaps.}

\item{schema_name}{Character. Target schema. This is optional, default is sex_ratio_evolution}

\item{url}{Character. OBM server URL (default: "https://openbiomaps.org").}

\item{project}{Character. OBM project name (default: "sex_ratio_evolution").}
}
\value{
Creates two Excel files in the working directory:
\itemize{
\item \code{table_metadata.xlsx}
\item \code{variable_metadata.xlsx}
}
}
\description{
Connects to an OpenBioMaps project using the obm.R package, retrieves data from a specified table,
and generates two Excel templates (\code{table_metadata.xlsx} and \code{variable_metadata.xlsx})
to help users document table- and variable-level metadata.
}
\examples{
\dontrun{
generate_obm_metadata("my_table", "optional: schema name")
}

}
